+++ src/geom.js
@@ -23,7 +23,7 @@ export function maxDistance3Rectangles(boxes, distance_fn) {
     const minx = Math.min(boxes[0].x1, boxes[1].x1, boxes[2].x1);
     const miny = Math.min(boxes[0].y1, boxes[1].y1, boxes[2].y1);
     const maxx = Math.max(boxes[0].x2, boxes[1].x2, boxes[2].x2);
-    const maxy = Math.max(boxes[0].y2, boxes[1].y2, boxes[2].y2);    
+    const maxy = Math.max(boxes[0].y2, boxes[1].y2, boxes[2].y2);
 
     let intersecting = false;
     for (let i = 0; i < 3; i++)
@@ -71,6 +71,22 @@ export function minDistance3Rectangles(boxes, distance_fn) {
                 const distance = distance_fn(i, j, k);
                 distanceMin = Math.min(distanceMin, distance);
             }
+
+    return distanceMin;
+}
+
+// Minimum possible distance between 2 rectangles
+// The proof can be deduced from Ondřej Palkovský's paper
+export function minDistance2Rectangles(boxes) {
+    const v0 = boxes[0].vertices();
+    const v1 = boxes[1].vertices();
+
+    let distanceMin = Infinity;
+    for (let i of v0)
+        for (let j of v1) {
+            const distance = i.distanceEarth(j);
+            distanceMin = Math.min(distanceMin, distance);
+        }
 
     return distanceMin;
 }
@@ -138,7 +154,7 @@ export function maxDistanceNRectangles(boxes) {
             boxes[i - 1].intersecting = true;
             boxes[i].intersecting = true;
         }
-    }    
+    }
 
     for (let i = 0; i < boxes.length; i++) {
         if (boxes[i].intersecting) {
@@ -223,6 +239,34 @@ export function findClosestPairIn2Segments(p1, p2, opt) {
     return min;
 }
 
+// Verify if there is a closing between range_a and range_b
+// TODO: Implement spatial caching
+function findClosestPairIn2PartialSegments(range_a, range_b, opt) {
+    const rtree = new Flatbush(range_a.end + 1 - range_a.start);
+    const lc = Math.abs(Math.cos(util.radians(opt.flight.flightPoints[range_a.start].y)));
+    for (let i = range_a.start; i <= range_a.end; i++) {
+        const r = opt.flight.flightPoints[i];
+        rtree.add(r.x * lc, r.y, r.x * lc, r.y);
+    }
+    rtree.finish();
+
+    let min = { d: Infinity };
+    for (let i = range_b.start; i <= range_b.end; i++) {
+        const pout = opt.flight.flightPoints[i];
+        const n = rtree.neighbors(pout.x * lc, pout.y, 1)[0] + range_a.start;
+        if (n !== undefined) {
+            const pin = opt.flight.flightPoints[n];
+            const d = pout.distanceEarth(pin);
+            if (d < min.d) {
+                min.d = d;
+                min.out = pout;
+                min.in = pin;
+            }
+        }
+    }
+    return min;
+}
+
 // Find the the furthest point between sega and segb from target
 // Exhaustive search with cache (O(n) worst case, O(log(n)) average)
 // The caching method works only when sega is the launch or segb is the landing
@@ -236,7 +280,7 @@ export function findFurthestPointInSegment(sega, segb, target, opt) {
         points = [target];
     else
         throw new TypeError('target must be either Point or Box');
-    
+
     let pos;
     let zSearch;
     if (sega === opt.launch) {
@@ -342,6 +386,15 @@ export function isTriangleClosed(p1, p2, distance, opt) {
             return f.o;
 
     const min = findClosestPairIn2Segments(p1, p2, opt);
+
+    if (min.d <= opt.scoring.closingDistance(distance, opt))
+        return min;
+    return false;
+}
+
+// Verify if there is a closing between sega and segb
+export function isOutAndReturnClosed(range_a, range_b, distance, opt) {
+    const min = findClosestPairIn2PartialSegments(range_a, range_b, opt);
 
     if (min.d <= opt.scoring.closingDistance(distance, opt))
         return min;